+++ CHANGELOG.rst
@@ -3,6 +3,20 @@ Change Log
 All notable changes to the COT project will be documented in this file.
 This project adheres to `Semantic Versioning`_.
 
+`1.5.1`_ - 2016-06-07
+---------------------
+
+**Added**
+
+- ``cot edit-hardware --network-descriptions`` option, to specify the
+  descriptive string(s) associated with each network definition.
+
+**Fixed**
+
+- `#48`_ - NIC type not set when adding NICs to an OVF that had none before.
+- When updating NIC network mapping, COT now also updates any Description
+  that references the network mapping.
+
 `1.5.0`_ - 2016-06-06
 ---------------------
 
@@ -385,6 +399,7 @@ Initial public release.
 .. _#44: https://github.com/glennmatthews/cot/issues/44
 .. _#45: https://github.com/glennmatthews/cot/issues/45
 .. _#47: https://github.com/glennmatthews/cot/issues/47
+.. _#48: https://github.com/glennmatthews/cot/issues/48
 
 .. _Semantic Versioning: http://semver.org/
 .. _`PEP 8`: https://www.python.org/dev/peps/pep-0008/
@@ -404,6 +419,7 @@ Initial public release.
 .. _argcomplete: https://argcomplete.readthedocs.io/en/latest/
 
 .. _Unreleased: https://github.com/glennmatthews/cot/compare/master...develop
+.. _1.5.1: https://github.com/glennmatthews/cot/compare/v1.5.0...v1.5.1
 .. _1.5.0: https://github.com/glennmatthews/cot/compare/v1.4.2...v1.5.0
 .. _1.4.2: https://github.com/glennmatthews/cot/compare/v1.4.1...v1.4.2
 .. _1.4.1: https://github.com/glennmatthews/cot/compare/v1.4.0...v1.4.1